Added Function: SendInputTap()
This function replaces the AutoIt MouseClick() function and sends input touch commands directly to the Memu instance using hidden cmd lines. This allows the mouse to be available for other uses, while the script runs. Initial implementation of this function was added to the Multi-Host script along with some additional screen conditions.