Added Function: MemoryReadPixel()
MemoryReadPixel() Function replaces PixelGetColor() as does not require the target window to be active in order to gather pixel information. This function is also much faster than PixelGetColor() when using the test case in the reference comment. PixelCheckColorBool() was reworked to use MemoryReadPixel().
Script Version updated to 0.2.0 as the main I/O functions are now in place.